create fft module
