+++ cryptography/polynomial/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod domain;
+mod fft;
 pub mod poly_coeff;
